first_chatbot/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
h11==0.14.0
haystack-ai==2.7.0
haystack-experimental==0.5.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
lazy_imports==0.4.0
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==10.6.0
networkx==3.4.2
numpy==2.2.2
openai==1.61.0
pandas==2.2.3
posthog==3.11.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0